@@ -29,13 +29,13 @@
     "@remix-run/node": "1.14.3",
     "@remix-run/react": "1.14.3",
     "@remix-run/vercel": "^1.14.3",
-    "@tabler/icons-react": "^2.11.0",
+    "@tabler/icons-react": "^2.12.0",
     "@tailwindcss/typography": "^0.5.9",
-    "@vercel/node": "^2.10.0",
+    "@vercel/node": "^2.10.1",
     "clsx": "^1.2.1",
     "daisyui": "^2.51.5",
     "dayjs": "^1.11.7",
-    "framer-motion": "^10.9.1",
+    "framer-motion": "^10.9.2",
     "gray-matter": "4.0.3",
     "isbot": "^3.6.7",
     "mermaid": "^10.0.2",
@@ -43,28 +43,28 @@
     "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "^18.2.0",
-    "tailwindcss": "^3.2.7"
+    "tailwindcss": "^3.3.0"
   },
   "devDependencies": {
-    "@nnecec/eslint-config": "^0.2.4",
-    "@nnecec/prettier-config": "^0.2.1",
+    "@nnecec/eslint-config": "^0.2.5",
+    "@nnecec/prettier-config": "^0.2.2",
     "@remix-run/dev": "1.14.3",
     "@remix-run/eslint-config": "1.14.3",
     "@remix-run/serve": "1.14.3",
     "@remix-run/server-runtime": "^1.14.3",
-    "@types/node": "^18.15.10",
+    "@types/node": "^18.15.11",
     "@types/prismjs": "^1.26.0",
-    "@types/react": "^18.0.30",
+    "@types/react": "^18.0.31",
     "@types/react-copy-to-clipboard": "^5.0.4",
     "@types/react-dom": "^18.0.11",
     "autoprefixer": "^10.4.14",
-    "eslint": "^8.36.0",
+    "eslint": "^8.37.0",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.21",
     "postcss-cli": "^10.1.0",
     "postcss-import": "^15.1.0",
     "prettier": "^2.8.7",
-    "tailwindcss": "^3.2.7",
+    "tailwindcss": "^3.3.0",
     "typescript": "^5.0.2"
   },
   "packageManager": "pnpm@7.16.0",
